fix: move score_render to lms
